Polya tree bugfixes.
SBST: make xa=false and xb=true when instantiating BernoulliProcess. This is needed as we code false for left branch and true for right branch, and exclusion coding assume leftbranch corresponds to 'xa' in BernoulliProcess.

BernoulliProcess: swap xa and xb in exclusion decoding. If det==num, this means P(X=xa)==1.